Add `opatch_patches` structured fact
Useful if you need to know what patches have already been applied.
I've done this as a `structured fact` as Puppet 3 is EOL at the end of
the year and soon everyone will be using the `facts` hash that is
enabled by default in puppet 4.

Example output...
```
[root@host ~]# facter -p opatch_patches
{
  /u01/app/oracle/product/12.1.0/dbhome_1 => [
    {
      patch_id => "21485069",
      patch_desc => ""
    },
    {
      patch_id => "21948354",
      patch_desc => "Database Patch Set Update : 12.1.0.2.160119 (21948354)"
    }
  ]
}
```
Originally developed for just oracle database, but seems to work fine with weblogic too.
```
[root@host ~]# facter -p opatch_patches
{
  /u01/app/oracle/Middleware => [
    {
      patch_id => "23094292",
      patch_desc => ""
    }
  ]
}
```

Signed-off-by: Alexander Fisher <alex@linfratech.co.uk>